+++ src/hardcopy.c
@@ -2769,8 +2769,8 @@ mch_print_begin(prt_settings_T *psettings)
 	// derive the bbox from that point.  We have the expected cpl chars
 	// across the media and lpp lines down the media.
 	bbox[1] = (int)(top - (psettings->lines_per_page + prt_header_height())
-							    * prt_line_height);
-	bbox[2] = (int)(left + psettings->chars_per_line * prt_char_width
+							    * (double)prt_line_height);
+	bbox[2] = (int)(left + psettings->chars_per_line * (double)prt_char_width
 									+ 0.5);
 	bbox[3] = (int)(top + 0.5);
     }
@@ -2782,7 +2782,7 @@ mch_print_begin(prt_settings_T *psettings)
 	bbox[1] = (int)bottom;
 	bbox[2] = (int)(left + ((psettings->lines_per_page
 			      + prt_header_height()) * prt_line_height) + 0.5);
-	bbox[3] = (int)(bottom + psettings->chars_per_line * prt_char_width
+	bbox[3] = (int)(bottom + psettings->chars_per_line * (double)prt_char_width
 									+ 0.5);
     }
     prt_dsc_ints("BoundingBox", 4, bbox);
